handle studentRepo contain name(like) Closes #62

+++ backend/src/main/java/com/spring/student/controller/StudentController.java
@@ -70,7 +70,7 @@ public class StudentController {
     }
 
     @GetMapping("/students/lengthName")
-    public int getStudentLengthByName(@RequestParam  String name){
+    public Long getStudentLengthByName(@RequestParam  String name){
         return studentService.getStudentLengthByName(name);
     }
 }
